VITIS-9944 Move platform reports into unique folder
