+++ src/runtime_src/core/tools/common/reports/platform/ReportRyzenPlatform.cpp
